Moved my homepage components to a new components/Home folder
